:arrow_up: Upgrading dependencies. #3216 MybatisPlus升级到3.4.3.4版本

+++ pigx-common/pigx-common-data/src/main/java/com/pig4cloud/pigx/common/data/datascope/SelectCountByScope.java
@@ -23,7 +23,7 @@ public class SelectCountByScope extends AbstractMethod {
 				this.sqlComment());
 		SqlSource sqlSource = languageDriver.createSqlSource(configuration, sql, modelClass);
 
-		return this.addSelectMappedStatementForOther(mapperClass, "selectCountByScope", sqlSource, Integer.class);
+		return this.addSelectMappedStatementForOther(mapperClass, "selectCountByScope", sqlSource, Long.class);
 	}
 
 }
